Add files via upload
This project brings a virtual drum kit to life! Click on the drum elements or use the keyboard to play different drum sounds, creating your own rhythm and beats. Built with HTML, CSS, and JavaScript, this interactive project is perfect for music lovers and front-end developers looking to have fun with web audio. Explore the code, experiment with sounds, and turn your keyboard into a mini drum set!